Predicate based testing in Java 7
An effort at predicate based testing in Java 7 with an LL(1) parser
implementation.

In conclusion I am pretty pleased with the implementation although,
unlike in langugaes that support higher-order functions and lambda
functions, this implementation needs to be dragged into a library
as it is tedious to build this every time.